daemon_set and stateful_set extract
